CryptoAgent/pyproject.toml
[project]
name = "cryptoagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.13",
    "flask>=3.1.1",
    "gradio>=5.34.2",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]
